tests/wpt/tests/navigation-timing/secure-connection-start-reuse.https.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Navigation Timing secureconnectionstart reuse</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
const t = async_test("Test that secureConnectionStart value is as expected when reused");
window.addEventListener("message", t.step_func_done(e => {
  assert_equals(e.data, "PASS");
}));
</script>
<iframe src="resources/secure-connection-interim.html">
